build(deps): update Spring Boot to 3.3.1 in dynamic-programming-demo
Update Spring Boot dependencies and related configurations in the
dynamic-programming-demo module to use version 3.3.1. This includesupdating the parent POM and dependency management section, as well as
adjusting the plugin configuration.

Additionally, update Hutool library from 5.8.21 to 5.8.25.

BREAKING CHANGE: This commit changes the base image for the Dockerfile
to use adoptopenjdk:11-jre-hotspot instead of openjdk:8-jre-alpine. This
could potentially break compatibility with systems or deployments that
rely on the previous Java version.